requirements.txt
argcomplete==1.8.1
asgiref==3.6.0
blinker==1.4
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
cliapp==1.20180812.1
cmdtest==0.32+git
command-not-found==0.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.5
ecdsa==0.18.0
filelock==3.6.0
fonttools==4.38.0
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.6.3
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.1
oauthlib==3.2.0
packaging==23.0
Pillow==9.4.0
platformdirs==2.5.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sdkmanager==0.5.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
ttystatus==0.38
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.13.0+ds
wadllib==1.3.6
zipp==1.0.0